Fix #21483 (part of #18384): Converted community-lessons to new progress tab for learner dashboard redesign (#21563)
* Renamed progress- tab references to old progress

* Renamed community-lesson tab references to progress

* Swapped community-lessons and progress i18n keys

* Moved async subtopic api call from new progress to learner-dashboard

* Passed subtopic mastery to new progress tab

* Removed async call from progress tab and updated test

* Renamed function to setTopicMastery

* Renamed variables for clarity

* Moved collections and explorations fetch to ngOnInit so can be used in Promise.all()

* Removed extra comments, getExplorationsAndCollection (moved to ngOnInit), subsciptions from new progress tab

* Deleted learnerdashboardbackendapi from new progress tests

* Updated progress-tab references to old-progress-tab

* Renamed progress-tab files to old-progress-tab

* Updated community-lessons tabs to progress-tabs files

* Fixed subtopic masteries test in new progress tab

* Renamed community-lessons to progress

* Fixed test

* Fixed test
