+++ core/templates/pages/learner-dashboard-page/learner-dashboard-page.module.ts
@@ -29,9 +29,9 @@ import {GoalListComponent} from './goal-list/goal-list.component';
 import {SkillCardModule} from './skill-card/skill-card.module';
 import {LearnerStorySummaryTileComponent} from 'components/summary-tile/learner-story-summary-tile.component';
 import {LearnerTopicGoalsSummaryTileComponent} from 'components/summary-tile/learner-topic-goals-summary-tile.component';
-import {ProgressTabComponent} from './progress-tab.component';
+import {OldProgressTabComponent} from './old-progress-tab.component';
 import {GoalsTabComponent} from './goals-tab.component';
-import {CommunityLessonsTabComponent} from './community-lessons-tab.component';
+import {ProgressTabComponent} from './progress-tab.component';
 import {LearnerTopicSummaryTileComponent} from 'components/summary-tile/learner-topic-summary-tile.component';
 import {HomeTabComponent} from './home-tab.component';
 import {LearnerGroupsTabComponent} from './learner-groups-tab.component';
@@ -85,12 +85,12 @@ import {NgCircleProgressModule} from 'ng-circle-progress';
     LearnerDashboardPageRootComponent,
     LearnerStorySummaryTileComponent,
     LearnerTopicGoalsSummaryTileComponent,
-    ProgressTabComponent,
+    OldProgressTabComponent,
     GoalsTabComponent,
     HomeTabComponent,
     LearnerGroupsTabComponent,
     LearnerTopicSummaryTileComponent,
-    CommunityLessonsTabComponent,
+    ProgressTabComponent,
     RemoveActivityModalComponent,
     LearnerDashboardSuggestionModalComponent,
     DeclineInvitationModalComponent,
@@ -106,12 +106,12 @@ import {NgCircleProgressModule} from 'ng-circle-progress';
     LearnerDashboardPageComponent,
     LearnerStorySummaryTileComponent,
     LearnerTopicGoalsSummaryTileComponent,
-    ProgressTabComponent,
+    OldProgressTabComponent,
     GoalsTabComponent,
     HomeTabComponent,
     LearnerGroupsTabComponent,
     LearnerTopicSummaryTileComponent,
-    CommunityLessonsTabComponent,
+    ProgressTabComponent,
     RemoveActivityModalComponent,
     LearnerDashboardSuggestionModalComponent,
     DeclineInvitationModalComponent,
