+++ core/templates/pages/learner-dashboard-page/progress-tab.component.css
